% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyyR.R
\name{abbyy_GET}
\alias{abbyy_GET}
\title{Base POST AND GET functions. Not exported.}
\usage{
abbyy_GET(path, query)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}
}
\value{
list
}
\description{
GET
}

