% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyyR.R
\name{abbyy_POST}
\alias{abbyy_POST}
\title{POST}
\usage{
abbyy_POST(path, query, body = "")
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{body}{passing image through body}
}
\value{
list
}
\description{
POST
}

