% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{deleteTask}
\alias{deleteTask}
\title{Delete Task}
\usage{
deleteTask(app_id = app_id, app_password = app_password, taskId)
}
\arguments{
\item{taskId}{}
}
\description{
This function deletes the given task and the images associated with this task.
}
\examples{
deleteTask(app_id=app_id, app_password=app_password, taskId)
}
\keyword{Delete}
\keyword{Task}

