% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{getAppInfo}
\alias{getAppInfo}
\title{Get Application Info}
\usage{
getAppInfo(app_id = app_id, app_password = app_password)
}
\arguments{
\item{app_id}{- get this from http://ocrsdk.com/. Set it before you use the package.}

\item{app_password}{- get this from http://ocrsdk.com/. Set it before you use the package.}
}
\description{
This function gets Information about a particular application
}
\examples{
getAppInfo(app_id=app_id, app_password=app_password)
}
\references{
\url{http://dhttp://ocrsdk.com/documentation/apireference/getApplicationInfo/}

\url{http://ocrsdk.com/schema/appInfo-1.0.xsd}
}
\keyword{Application}
\keyword{Information}

