% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{getResults}
\alias{getResults}
\title{Get Results}
\usage{
getResults(output = NULL)
}
\arguments{
\item{output}{Optional; folder to which you want to save the data from the processed images}
}
\description{
Get data from all the processed images.
The function prints the status of the task by default.
The function returns a data.frame with all the task details: id (task id), registrationTime, statusChangeTime, status (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), filesCount (No. of files), credits, resultUrl (URL for the processed file if applicable)
}
\examples{
# getTaskStatus(taskId="task_id")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getTaskStatus/}
}
\keyword{Get}
\keyword{Results}

