% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{getTaskStatus}
\alias{getTaskStatus}
\title{Get Task Status}
\usage{
getTaskStatus(taskId = NULL)
}
\arguments{
\item{taskId}{}
}
\description{
This function gets task status for a particular task id
}
\examples{
getTaskStatus(taskId="task_id")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getTaskStatus/}
}
\keyword{Status}
\keyword{Task}

