% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{listTasks}
\alias{listTasks}
\title{List Tasks}
\usage{
listTasks(fromDate = NULL, toDate = NULL, excludeDeleted = "false")
}
\arguments{
\item{fromDate;}{not required;  format: yyyy-mm-ddThh:mm:ssZ}

\item{toDate;}{not required;  format: yyyy-mm-ddThh:mm:ssZ}

\item{excludeDeleted;}{not required; default='false'}
}
\description{
This function gets Information about a particular application
}
\examples{
listTasks(fromDate=NULL,toDate=NULL,excludeDeleted='false')
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}
}
\keyword{List}
\keyword{Tasks}

