% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocr_file.R
\name{ocrFile}
\alias{ocrFile}
\title{OCR File}
\usage{
ocrFile(file_path = NULL, output_dir = "./", exportFormat = "txt")
}
\arguments{
\item{output_dir}{path to output directory. file_name will be same as input file name (except for the extension)}

\item{exportFormat}{optional, default: txt;
options: txt, txtUnstructured, rtf, docx, xlsx, pptx, pdfSearchable, pdfTextAndImages, pdfa, xml, xmlForCorrectedImage, alto}

\item{path_to_ocr}{path to file containing OCR'd text; required}
}
\value{
path to output file
}
\description{
Want to quick OCR a local file and get the results? Use this function.
}
\examples{
\dontrun{
OCRFile(file_path="path_to_ocr_file", output_dir="path_to_output_dir")
}
}

