% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processBusinessCard}
\alias{processBusinessCard}
\title{Process Business Card}
\usage{
processBusinessCard(file_path = NULL, language = "English",
  profile = "documentConversion", textType = "normal",
  imageSource = "auto", correctOrientation = "true", correctSkew = "true",
  readBarcodes = "false", exportFormat = "txt", description = "",
  pdfPassword = "")
}
\arguments{
\item{file_path:}{path of the document}
}
\description{
This function gets Information about a particular application
}
\examples{
processBusinessCard(file_path="file_path", language="English", profile="documentConversion",textType="normal", imageSource="auto", correctOrientation="true", correctSkew="true", readBarcodes,exportFormat="txt",description="", pdfPassword="")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processBusinessCard/}
}
\keyword{Business}
\keyword{Card}

