% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processCheckmarkField}
\alias{processCheckmarkField}
\title{processCheckmarkField Method}
\usage{
processCheckmarkField(file_path = NULL)
}
\arguments{
\item{app_id}{- get this from http://ocrsdk.com/. Set it before you use the package.}

\item{app_password}{- get this from http://ocrsdk.com/. Set it before you use the package.}
}
\description{
This function gets Information about a particular application
}
\examples{
processCheckmarkField()
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processCheckmarkField/}
}
\keyword{Application}
\keyword{Information}

