% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processDocument}
\alias{processDocument}
\title{Process Document}
\usage{
processDocument()
}
\arguments{
\item{app_id}{- get this from http://ocrsdk.com/. Set it before you use the package.}

\item{app_password}{- get this from http://ocrsdk.com/. Set it before you use the package.}
}
\description{
This function processes several images
}
\examples{
processDocument()
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processDocument/}
}
\keyword{Application}
\keyword{Information}

