% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processFields}
\alias{processFields}
\title{Process Fields}
\usage{
processFields(file_path = NULL, taskId = NULL, description = "")
}
\arguments{
\item{file_path}{path of the document}

\item{taskId}{- Only tasks with Submitted, Completed or NotEnoughCredits status can be processed using this function.}
}
\description{
This function gets Information about a particular application
}
\examples{
processFields(file_path="file_path", taskId="task_id",description="")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processFields/}
}
\keyword{Application}
\keyword{Information}

