% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processImage}
\alias{processImage}
\title{Process Image}
\usage{
processImage(file_path = NULL, language = "English",
  profile = "documentConversion", textType = "normal",
  imageSource = "auto", correctOrientation = "true", correctSkew = "true",
  readBarcodes = "false", exportFormat = "txt", description = "",
  pdfPassword = "")
}
\arguments{
\item{language;}{optional, default: English}

\item{profile;}{optional, default: documentConversion}

\item{textType;}{optional, default: normal}

\item{imageSource;}{optional, default: auto}

\item{correctOrientation;}{optional, default: true}

\item{correctSkew;}{optional, default: true}

\item{readBarcodes;}{optional, default:}

\item{exportFormat;}{optional, default: txt}

\item{pdfPassword;}{optional, default: NULL}

\item{description;}{optional, default: ""}
}
\description{
This function processes an image
}
\examples{
processImage(language="English", profile="documentConversion",textType="normal", imageSource="auto", correctOrientation="true", correctSkew="true", readBarcodes,exportFormat="txt",description="", pdfPassword="")
}
\references{
\url{http://ocrsdk.com/documentation/specifications/image-formats/}

\url{http://ocrsdk.com/documentation/apireference/processImage/}
}
\keyword{Application}
\keyword{Information}

