% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processMRZ}
\alias{processMRZ}
\title{Process MRZ: Extract data from Machine Readable Zone}
\usage{
processMRZ(file_path = NULL)
}
\arguments{
\item{file_path}{path of the document
Extract data from Machine Readable Zone in an Image}

\item{file_path}{path to the document}
}
\description{
Process MRZ: Extract data from Machine Readable Zone
}
\examples{
processMRZ(file_path="file_path")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processMRZ/}
}
\keyword{Machine}
\keyword{Readable}
\keyword{Zone}

