% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setapp.R
\name{setapp}
\alias{setapp}
\title{Sets Application ID and Password}
\usage{
setapp(appdetails = NULL)
}
\arguments{
\item{appdetails}{Required; A vector of app_id, app_password. Get these from \url{http://ocrsdk.com/}. Set them before you use other functions.}
}
\description{
Set Application ID and Password. Needed for interfacing with Abbyy FineReader Cloud OCR SDK. Run this before anything else.
}
\examples{
\dontrun{
setapp(c("app_id", "app_password"))
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}
}
\keyword{Application}
\keyword{ID}
\keyword{Password}
\keyword{Sets}
\keyword{and}

