% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{submitImage}
\alias{submitImage}
\title{Submit Image}
\usage{
submitImage(file_path)
}
\arguments{
\item{taskId}{- Optional.}

\item{pdfPassword}{- Optional. If the pdf is password protected, put the password here.}
}
\description{
Adds image to the existing task or creates a new task for the uploaded image. The new task isn't processed till processDocument or processFields is called.
}
\examples{
submitImage(file_path="/images/image1.png",taskId="task_id",pdfPassword="pdf_password")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/submitImage/}
}
\keyword{Image}
\keyword{Submit}

