% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyyR.R
\name{abbyy_check}
\alias{abbyy_check}
\title{Request Response Verification}
\usage{
abbyy_check(req)
}
\arguments{
\item{req}{request}
}
\value{
in case of failure, a message
}
\description{
Request Response Verification
}

