% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\docType{package}
\name{abbyyR-package}
\alias{abbyyR}
\alias{abbyyR-package}
\title{abbyyR: R Client for the Abbyy Cloud OCR}
\description{
Easily OCR images, barcodes, forms, documents with machine readable zones, e.g. passports, right from R.
Get the results in form of text files or detailed XML. The package provides access to Abbyy OCR -- see \url{http://ocrsdk.com/}.
Details about results of calls to the API can be found at \url{http://ocrsdk.com/documentation/specifications/status-codes/}.

Your need credentials (application ID and password) to use this application.
If you haven't already, you can get this at \url{http://ocrsdk.com/}.
}
\author{
Gaurav Sood
}

