% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{deleteTask}
\alias{deleteTask}
\title{Delete Task}
\usage{
deleteTask(taskId = NULL)
}
\arguments{
\item{taskId}{}
}
\description{
This function deletes a particular task and associated data. From Abbyy "If you try to delete the task that has already been deleted, the successful response is returned."
The function by default prints the status of the task you are trying to delete. It will show up as 'deleted' if successful
The function returns a data.frame with all the details of the task you are trying to delete: id (task id), registrationTime, statusChangeTime, status (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), filesCount (No. of files), credits, resultUrl (URL for the processed file if applicable)
}
\examples{
deleteTask(taskId="task_id")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/deleteTask/}
}
\keyword{Delete}
\keyword{Task}

