% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAppInfo.R
\name{getAppInfo}
\alias{getAppInfo}
\title{Get Application Info}
\usage{
getAppInfo(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{abbyy_GET}}.}
}
\value{
A data.frame with the following fields: Name of the Application (name), 
No. of pages remaining (pages), No. of fields remaining (fields), 
when the application credits expire (expires) and type of application (type).
}
\description{
Get Information about the Application, such as number of pages (given the money), 
when the application credits expire etc. The function prints the details, and returns a 
data.frame with the details.
}
\examples{
\dontrun{
getAppInfo()
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}

\url{http://ocrsdk.com/schema/appInfo-1.0.xsd}
}

