% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAppInfo.R
\name{getAppInfo}
\alias{getAppInfo}
\title{Get Application Info}
\usage{
getAppInfo()
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), No. of fields remaining (given the money), and when the application credits expire.
}
\description{
Get Information about the Application, such as number of pages (given the money), when the application credits expire etc. The function prints the details, and returns a list with the details.
}
\examples{
\dontrun{
getAppInfo()
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}

\url{http://ocrsdk.com/schema/appInfo-1.0.xsd}
}

