% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{getAppInfo}
\alias{getAppInfo}
\title{Get Application Info}
\usage{
getAppInfo()
}
\description{
Get Information about the Application including details like: Name of the Application, No. of pages remaining (given the money), No. of fields remaining (given the money), and when the application credits expire. The function automatically prints these out. It also stores these in a list.
}
\examples{
# getAppInfo()
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}

\url{http://ocrsdk.com/schema/appInfo-1.0.xsd}
}
\keyword{Application}
\keyword{Get}
\keyword{Information}

