% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Get Results}
\usage{
getResults(output="", save_to_file=TRUE)
}
\arguments{
\item{output}{Optional; folder to which you want to save the data from the processed images; Default is same folder as the script}

\item{save_to_file}{Required; Default is TRUE; If true, it saves to file. Otherwise returns a data frame with results + other attributes from Abbyy}
}
\description{
Get data from all the processed images.
The function goes through the finishedTasks data frame and downloads all the files in resultsUrl
}
\examples{
\dontrun{
getResults(save_to_file=FALSE)
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getTaskStatus/}
}

