% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Get Results}
\usage{
getResults(output="")
}
\arguments{
\item{output}{Optional; folder to which you want to save the data from the processed images; Default is same folder as the script}
}
\description{
Get data from all the processed images.
The function goes through the finishedTasks data frame and downloads all the files in resultsUrl
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getTaskStatus/}
}

