% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{getTaskStatus}
\alias{getTaskStatus}
\title{Get Task Status}
\usage{
getTaskStatus(app_id = app_id, app_password = app_password, taskId)
}
\arguments{
\item{taskId}{}
}
\description{
This function gets Information about a particular application
}
\examples{
getTaskStatus(app_id=app_id, app_password=app_password, taskId)
}
\keyword{Status}
\keyword{Task}

