% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaskStatus.R
\name{getTaskStatus}
\alias{getTaskStatus}
\title{Get Task Status}
\usage{
getTaskStatus(taskId = NULL, ...)
}
\arguments{
\item{taskId}{Required, Id of the task}

\item{\dots}{Additional arguments passed to \code{\link{abbyy_GET}}.}
}
\value{
A \code{data.frame} with all the available details about the task
}
\description{
This function gets task status for a particular task ID.
The function prints the status of the task by default.
The function returns a data.frame with all the task details: id (task id), registrationTime, 
statusChangeTime, status (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), 
filesCount (No. of files), credits, resultUrl (URL for the processed file if applicable)
}
\examples{
\dontrun{
getTaskStatus(taskId="task_id")
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getTaskStatus/}
}
