% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{getTaskStatus}
\alias{getTaskStatus}
\title{Get Task Status}
\usage{
getTaskStatus(taskId = NULL)
}
\arguments{
\item{taskId}{-id of the task; required}
}
\value{
A data frame with all the available details about the task
}
\description{
This function gets task status for a particular task ID.
The function prints the status of the task by default.
The function returns a data.frame with all the task details: id (task id), registrationTime, statusChangeTime, status (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), filesCount (No. of files), credits, resultUrl (URL for the processed file if applicable)
}
\examples{
# getTaskStatus(taskId="task_id")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getTaskStatus/}
}

