% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listFinishedTasks.R
\name{listFinishedTasks}
\alias{listFinishedTasks}
\title{List Finished Tasks}
\usage{
listFinishedTasks()
}
\value{
A data frame with the following columns: id (task id), registrationTime, statusChangeTime, status (Completed), filesCount (No. of files), credits, resultUrl (URL for the processed file)
}
\description{
List all the finished tasks in the application.
From Abbyy FineReader: The tasks are ordered by the time of the end of processing. No more than 100 tasks can be returned at one method call.
The function prints number of finished tasks by default
}
\references{
\url{http://ocrsdk.com/documentation/apireference/listFinishedTasks/}
}

