% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{listFinishedTasks}
\alias{listFinishedTasks}
\title{List Finished Tasks}
\usage{
listFinishedTasks(app_id = app_id, app_password = app_password)
}
\description{
This function gets Information about a particular application
}
\examples{
listFinishedTasks(app_id=app_id, app_password=app_password)
}
\keyword{Finished}
\keyword{Tasks}

