% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{listFinishedTasks}
\alias{listFinishedTasks}
\title{List Finished Tasks}
\usage{
listFinishedTasks()
}
\description{
This function gets Information about a particular application
}
\examples{
listFinishedTasks()
}
\references{
\url{http://ocrsdk.com/documentation/apireference/listFinishedTasks/}
}
\keyword{Finished}
\keyword{Tasks}

