% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listTasks.R
\name{listTasks}
\alias{listTasks}
\title{List Tasks}
\usage{
listTasks(fromDate=NULL,toDate=NULL,excludeDeleted='false')
}
\arguments{
\item{fromDate}{- optional;  format: yyyy-mm-ddThh:mm:ssZ}

\item{toDate}{- optional;  format: yyyy-mm-ddThh:mm:ssZ}

\item{excludeDeleted}{- optional; default='false'}
}
\value{
A data frame with the following columns: id (task id), registrationTime, statusChangeTime, status
		   (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), filesCount (No. of files), credits, resultUrl (URL for the processed file)
}
\description{
List all the tasks in the application. You can specify a date range and whether or not you want to include deleted tasks.
The function prints Total number of tasks and No. of Finished Tasks.
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}
}

