% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{listTasks}
\alias{listTasks}
\title{List Tasks}
\usage{
listTasks(appId = app_id, appPass = app_password, fromDate, toDate,
  excludeDeleted)
}
\arguments{
\item{fromDate;}{not required;  format: yyyy-mm-ddThh:mm:ssZ}

\item{toDate;}{not required;  format: yyyy-mm-ddThh:mm:ssZ}

\item{excludeDeleted;}{not required; default='false'}
}
\description{
This function gets Information about a particular application
}
\examples{
listTasks(app_id=app_id, app_password=app_password,fromDate,toDate,excludeDeleted)
}
\keyword{List}
\keyword{Tasks}

