% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{listTasks}
\alias{listTasks}
\title{List Tasks}
\usage{
listTasks(fromDate = NULL, toDate = NULL, excludeDeleted = "false")
}
\arguments{
\item{fromDate}{- optional;  format: yyyy-mm-ddThh:mm:ssZ}

\item{toDate}{- optional;  format: yyyy-mm-ddThh:mm:ssZ}

\item{excludeDeleted}{- optional; default='false'}
}
\description{
List all the tasks in the application. You can specify a date range and whether or not you want to include deleted tasks.
The function prints Total number of tasks, Task IDs, and No. of Finished Tasks.
The function returns a data.frame with the following columns: id (task id), registrationTime, statusChangeTime, status (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), filesCount (No. of files), credits, resultUrl (URL for the processed file)
}
\examples{
# listTasks(fromDate=NULL,toDate=NULL,excludeDeleted='false')
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}
}
\keyword{List}
\keyword{Tasks}

