% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processBarcodeField}
\alias{processBarcodeField}
\title{Process Bar Code Field}
\usage{
processBarcodeField(file_path = NULL)
}
\arguments{
\item{file_path:}{path of the document}
}
\description{
This function gets Information about a particular application
}
\examples{
processBarcodeField(file_path="file_path")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processBarcodeField/}
}
\keyword{Application}
\keyword{Information}

