% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBusinessCard.R
\name{processBusinessCard}
\alias{processBusinessCard}
\title{Process Business Card}
\usage{
processBusinessCard(file_path = "", language = "English",
  imageSource = "auto", correctOrientation = "true",
  correctSkew = "true", exportFormat = "vCard", description = "",
  pdfPassword = "", ...)
}
\arguments{
\item{file_path}{required, path of the document, default: ""}

\item{language}{optional, default: English}

\item{imageSource}{optional, default: auto}

\item{correctOrientation}{optional, default: true}

\item{correctSkew}{optional, default: true}

\item{exportFormat}{optional, default: "vCard"}

\item{description}{optional, default: ""}

\item{pdfPassword}{optional, default: NULL}

\item{\dots}{Additional arguments passed to \code{\link{abbyy_POST}}.}
}
\value{
Data frame with details of the task associated with the submitted Business Card
}
\description{
Processes a Business Card
}
\examples{
\dontrun{
processBusinessCard(file_path="file_path", language="English")
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processBusinessCard/}
}
\keyword{Image}
\keyword{Process}
\keyword{Remote}
