% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processBusinessCard}
\alias{processBusinessCard}
\title{Process Business Card}
\usage{
processBusinessCard(file_path = NULL, language = "English",
  imageSource = "auto", correctOrientation = "true", correctSkew = "true",
  exportFormat = "vCard", description = "", pdfPassword = "")
}
\arguments{
\item{file_path}{path of the document}

\item{language}{optional, default: English}

\item{imageSource}{optional, default: auto}

\item{correctOrientation}{optional, default: true}

\item{correctSkew}{optional, default: true}

\item{exportFormat}{optional, default: "vCard"}

\item{description}{optional, default: ""}

\item{pdfPassword}{optional, default: NULL}
}
\value{
Data frame with details of the task associated with the submitted Business Card
}
\description{
This function gets Information about a particular application
}
\examples{
# processBusinessCard(file_path="file_path", language="English")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processBusinessCard/}
}
\keyword{Image}
\keyword{Process}
\keyword{Remote}

