% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCheckmarkField.R
\name{processCheckmarkField}
\alias{processCheckmarkField}
\title{processCheckmarkField Method}
\usage{
processCheckmarkField(file_path = "", checkmarkType = "empty",
  region = "-1,-1,-1,-1", correctionAllowed = "false", pdfPassword = "",
  description = "")
}
\arguments{
\item{file_path}{required, path of the document, default: ""}

\item{checkmarkType}{optional, default: "empty"}

\item{region}{coordinates of region from top left, 4 values: top left bottom right; optional; default: "-1,-1,-1,-1" (entire image)}

\item{correctionAllowed}{optional, default: "false"}

\item{pdfPassword}{optional, default: ""}

\item{description}{optional, default: ""}
}
\value{
Data frame with details of the task associated with the submitted Image
}
\description{
This function gets Information about a particular application
}
\examples{
\dontrun{
processCheckmarkField(file_path="file_path")
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processCheckmarkField/}
}

