% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processDocument}
\alias{processDocument}
\title{Process Document}
\usage{
processDocument(taskId = NULL)
}
\arguments{
\item{taskId}{- Only tasks with Submitted, Completed or NotEnoughCredits status can be processed using this function.}
}
\description{
This function processes several images for the same task and results in a multi-page document.
For instance, upload pages of the book individually via submitImage to the same task. And then process it via ProcessDocument to get a multi-page pdf.
}
\examples{
# processDocument(taskId = "task_id")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processDocument/}
}
\keyword{Document}
\keyword{Process}

