% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processImage}
\alias{processImage}
\title{Process Image}
\usage{
processImage(file_path = NULL, language = "English",
  profile = "documentConversion", textType = "normal",
  imageSource = "auto", correctOrientation = "true", correctSkew = "true",
  readBarcodes = "false", exportFormat = "txt", description = "",
  pdfPassword = "")
}
\arguments{
\item{file_path}{path of the document}

\item{language}{optional, default: English}

\item{profile}{optional, default: documentConversion}

\item{textType}{optional, default: normal}

\item{imageSource}{optional, default: auto}

\item{correctOrientation}{optional, default: true}

\item{correctSkew}{optional, default: true}

\item{readBarcodes}{optional, default:}

\item{exportFormat}{optional, default: txt}

\item{description}{optional, default: ""}

\item{pdfPassword}{optional, default: NULL}
}
\description{
This function processes an image
}
\examples{
# processImage(file_path="file_path", language="English")
}
\references{
\url{http://ocrsdk.com/documentation/specifications/image-formats/}

\url{http://ocrsdk.com/documentation/apireference/processImage/}
}
\keyword{Image}
\keyword{Process}

