% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processImage}
\alias{processImage}
\title{Process Image}
\usage{
processImage(app_id = app_id, app_password = app_password,
  language = "English", profile = "documentConversion",
  textType = "normal", imageSource = "auto", correctOrientation = "true",
  correctSkew = "true", readBarcodes, exportFormat = "txt",
  description = "", pdfPassword = "", file_path)
}
\arguments{
\item{language;}{default: English}

\item{profile;}{default: documentConversion}

\item{textType;}{default: normal}

\item{imageSource;}{default: auto}

\item{correctOrientation;}{default: true}

\item{correctSkew;}{default: true}

\item{readBarcodes;}{default:}

\item{exportFormat;}{default: txt}

\item{pdfPassword;}{default:}

\item{description;}{default:}
}
\description{
This function processes an image
}
\examples{
processImage(language="English", profile="documentConversion",textType="normal", imageSource="auto", correctOrientation="true", correctSkew="true",readBarcodes,exportFormat="txt",description="", pdfPassword="", file_path="file_path")
}
\references{
\url{http://dhttp://ocrsdk.com/documentation/apireference/processImage/}
}
\keyword{Application}
\keyword{Information}

