% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processMRZ}
\alias{processMRZ}
\title{Process MRZ}
\usage{
processMRZ(file_path = NULL)
}
\arguments{
\item{app_id}{- get this from http://ocrsdk.com/. Set it before you use the package.}

\item{app_password}{- get this from http://ocrsdk.com/. Set it before you use the package.}
}
\description{
This function gets Information about a particular application
}
\examples{
processMRZ()
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processMRZ/}
}
\keyword{Application}
\keyword{Information}

