% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMRZ.R
\name{processMRZ}
\alias{processMRZ}
\title{Process MRZ: Extract data from Machine Readable Zone}
\usage{
processMRZ(file_path = "", ...)
}
\arguments{
\item{file_path}{path to the document}

\item{\dots}{Additional arguments passed to \code{\link{abbyy_POST}}.}
}
\value{
Data frame with details of the task associated with the submitted MRZ document
}
\description{
Extract data from Machine Readable Zone in an Image
}
\examples{
\dontrun{
processMRZ(file_path="file_path")
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processMRZ/}
}
