% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processPhotoId}
\alias{processPhotoId}
\title{Process Photo ID}
\usage{
processPhotoId(file_path = NULL)
}
\arguments{
\item{app_id}{- get this from http://ocrsdk.com/. Set it before you use the package.}

\item{app_password}{- get this from http://ocrsdk.com/. Set it before you use the package.}
}
\description{
This function gets Information about a particular application
}
\examples{
processPhotoId()
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processPhotoId/}
}
\keyword{Application}
\keyword{Information}

