% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processPhotoId}
\alias{processPhotoId}
\title{Process Photo ID}
\usage{
processPhotoId(file_path = NULL, idType = "auto", imageSource = "auto",
  correctOrientation = "true", correctSkew = "true", description = "",
  pdfPassword = "")
}
\arguments{
\item{file_path}{path to file; required}

\item{idType}{optional; default = "auto"}

\item{imageSource}{optional; default = "auto"}

\item{correctOrientation}{optional; default = "true"}

\item{description}{optional; default = ""}

\item{pdfPassword}{optional; default = ""}

\item{correctSke}{optional; default = "true"}
}
\description{
This function gets Information about a particular application
}
\examples{
processPhotoId(file_path="file_path", idType="auto", imageSource="auto", correctOrientation="true", correctSkew="true", description="", pdfPassword="")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processPhotoId/}
}
\keyword{ID}
\keyword{Photo}

