% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPhotoId.R
\name{processPhotoId}
\alias{processPhotoId}
\title{Process Photo ID}
\usage{
processPhotoId(file_path = "", idType = "auto", imageSource = "auto",
  correctOrientation = "true", correctSkew = "true", description = "",
  pdfPassword = "", ...)
}
\arguments{
\item{file_path}{path to file; required}

\item{idType}{optional; default = "auto"}

\item{imageSource}{optional; default = "auto"}

\item{correctOrientation}{optional; default = "true"}

\item{correctSkew}{optional; default = "true"}

\item{description}{optional; default = ""}

\item{pdfPassword}{optional; default = ""}

\item{\dots}{Additional arguments passed to \code{\link{abbyy_POST}}.}
}
\value{
Data frame with details of the task associated with the submitted Photo ID image
}
\description{
Get data from a Photo ID. The function is under testing and may not work fully.
}
\examples{
\dontrun{
processPhotoId(file_path="file_path", idType="auto", imageSource="auto")
}
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processPhotoId/}
}
