% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processRemoteImage}
\alias{processRemoteImage}
\title{Process Remote Image}
\usage{
processRemoteImage(img_url = NULL, language = "English",
  profile = "documentConversion", textType = "normal",
  imageSource = "auto", correctOrientation = "true", correctSkew = "true",
  readBarcodes = "false", exportFormat = "txt", description = "",
  pdfPassword = "")
}
\arguments{
\item{img_url}{url to remote image}

\item{language}{optional, default: English}

\item{profile}{optional, default: documentConversion}

\item{textType}{optional, default: normal}

\item{imageSource}{optional, default: auto}

\item{correctOrientation}{optional, default: true}

\item{correctSkew}{optional, default: true}

\item{readBarcodes}{optional, default:}

\item{exportFormat}{optional, default: txt}

\item{description}{optional, default: ""}

\item{pdfPassword}{optional, default: NULL}
}
\description{
This function gets Information about a particular application
}
\examples{
# processRemoteImage(img_url="img_url")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processRemoteImage/}
}
\keyword{Image}
\keyword{Process}
\keyword{Remote}

