% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processRemoteImage}
\alias{processRemoteImage}
\title{Process Remote Image}
\usage{
processRemoteImage(app_id = app_id, app_password = app_password, img_url)
}
\arguments{
\item{app_id}{- get this from http://ocrsdk.com/. Set it before you use the package.}

\item{app_password}{- get this from http://ocrsdk.com/. Set it before you use the package.}
}
\description{
This function gets Information about a particular application
}
\examples{
getAppInfo(img_url)
}
\keyword{Application}
\keyword{Information}

