% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{processTextField}
\alias{processTextField}
\title{Process Text Field}
\usage{
processTextField(file_path = NULL)
}
\arguments{
\item{file_path:}{path of the document}
}
\description{
This function gets Information about a particular application
}
\examples{
processTextField(file_path="file_path")
}
\references{
\url{http://ocrsdk.com/documentation/apireference/processTextField/}
}
\keyword{Field}
\keyword{Text}

