% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abbyy.R
\name{setapp}
\alias{setapp}
\title{Sets Application ID and Password}
\usage{
setapp(appdetails = NULL)
}
\arguments{
\item{appdetails:}{a vector of app_id, app_password. Get these from http://ocrsdk.com/. Set them before you use other functions.}
}
\description{
Sets Application ID and Password. Needed for interfacing with Abbyy FineReader Cloud OCR SDK
}
\examples{
setapp(c("app_id", "app_password"))
}
\references{
\url{http://ocrsdk.com/documentation/apireference/getApplicationInfo/}
}
\keyword{Application}
\keyword{ID}
\keyword{Password}
\keyword{Sets}
\keyword{and}

